Updates for RenderNode and a few other things

+++ RenderNode.py
@@ -28,11 +28,10 @@ import Utilities.TaskUtils as TaskUtils
 
 class RenderTCPServer(TCPServer):
     def __init__(self, *arglist, **kwargs):
-        #Check for another instance of RenderNodeMain.exe
         inst = checkRenderNodeInstances()
         if not inst:
             sys.exit(1)
-        #Initiate TCP Server
+
         self.renderServ = TCPServer.__init__(self, *arglist, **kwargs)
         if sys.platform == "win32":
             self.si = subprocess.STARTUPINFO()
@@ -42,12 +41,10 @@ class RenderTCPServer(TCPServer):
         self.rsGPUs = self.rsGPUs.split(",")[:-1]
         self.rsGPUids = [x.split(":")[0] for x in self.rsGPUs]
         if len(self.rsGPUs) != len(self.rsGPUids):
-            logger.error("Problems parsing RS Preferences")
-            raise Exception("Problems parsing RS Preferences")
-        logger.info("{0} Redshift Enabled GPU(s) found on this node".format(len(self.rsGPUs)))
-        logger.debug("GPUs available for rendering are {0}".format(self.rsGPUs))
+            logger.warning("Problems parsing RS Preferences")
+            logger.info("{0} Redshift Enabled GPU(s) found on this node".format(len(self.rsGPUs)))
+            logger.debug("GPUs available for rendering are {0}".format(self.rsGPUs))
 
-        #Setup class variables
         execs = hydra_executable.fetch()
         self.execsDict = {ex.name: ex.path for ex in execs}
 
@@ -57,27 +54,20 @@ class RenderTCPServer(TCPServer):
         self.thisNodeName = Utils.myHostName()
 
         #Cleanup job if we start with it assigned to us (Like if the node crashed/restarted)
-        logger.debug("Housekeeping...")
         thisNode = hydra_rendernode.fetch("WHERE host = %s", (self.thisNodeName,))
         query = "UPDATE hydra_rendernode SET status = %s WHERE host = %s"
         if thisNode.task_id:
             logger.warning("Rouge task discovered. Unsticking...")
             task = hydra_taskboard.fetch("WHERE id = %s", (thisNode.task_id,))
-            if thisNode.status == PENDING or thisNode.status == OFFLINE:
-                newStatus = OFFLINE
-            else:
-                newStatus = IDLE
-            TaskUtils.unstick(taskID=thisNode.task_id, newTaskStatus=CRASHED,
-                              host=thisNode.host, newHostStatus=newStatus)
+            newStatus = OFFLINE if thisNode.status in [OFFLINE, PENDING] else ONLINE
+            TaskUtils.unstickTask(taskID = thisNode.task_id, newTaskStatus = CRASHED,
+                              host = thisNode.host, newHostStatus = newStatus)
             JobUtils.manageNodeLimit(task.job_id)
-        elif thisNode.status == STARTED and not thisNode.task_id:
-            logger.warning("Reseting bad status.")
+        elif thisNode.status in [STARTED, PENDING] and not thisNode.task_id:
+            logger.warning("Reseting bad status, node set {0} but no task found!".format(thisNode.status))
+            newStatus = READY if thisNode.status == STARTED else OFFLINE
             with transaction() as t:
-                t.cur.execute(query, (READY, self.thisNodeName,))
-        elif thisNode.status == PENDING and not thisNode.task_id:
-            logger.warning("Reseting bad status.")
-            with transaction() as t:
-                t.cur.execute(query, (OFFLINE, self.thisNodeName,))
+                t.cur.execute(query, (newStatus, self.thisNodeName,))
 
         #Update current software version on the DB if necessary
         current_version = sys.argv[0]
@@ -91,23 +81,21 @@ class RenderTCPServer(TCPServer):
 
     def processRenderTasks(self):
         """The loop that looks for jobs on the DB and runs them if the node meets
-        the job's requirements (Priority & Capabilities)"""
+        the job's requirements"""
         thisNode = hydra_rendernode.fetch("WHERE host = %s",(self.thisNodeName,))
-        debugStr = "Host: {0} Status: {1} Capabilities {2}"
-        logger.debug(debugStr.format(thisNode.host,
-                                    niceNames[thisNode.status],
+        debugStr = "Host: {0} Status: {1} Capabilities: {2}"
+        logger.debug(debugStr.format(thisNode.host, niceNames[thisNode.status],
                                     thisNode.capabilities))
 
         #If this node is not idle, don't try to find a new job
         if thisNode.status != IDLE:
             return
 
-
         #Otherwise, get a job that's:
         #-Ready to be run and
         #-Has a high enough priority level for this particular node and
         #-Is able to meet to jobs required capabilities
-        whereClause = "WHERE status = %s AND priority = %s AND %s LIKE requirements AND archived = 0 AND failures NOT LIKE '%%{0}%%'".format(thisNode.host)
+        whereClause = "WHERE status = %s AND priority >= %s AND %s LIKE requirements AND archived = 0 AND failures NOT LIKE '%%{0}%%'".format(thisNode.host)
         whereTuple = (READY, thisNode.minPriority, thisNode.capabilities)
         orderTuples = (("priority", "DESC"), ("id", "ASC"))
 
@@ -117,8 +105,11 @@ class RenderTCPServer(TCPServer):
                                                         orderTuples = orderTuples,
                                                         limit = 1,
                                                         explicitTransaction = t)
+            #If not task is found, stop and wait to search again
             if not render_task:
                 return
+
+            #Otherwise, render the task
             render_job = hydra_jobboard.fetch("WHERE id = %s",
                                                 (render_task.job_id,),
                                                 explicitTransaction = t)
@@ -154,14 +145,8 @@ class RenderTCPServer(TCPServer):
             Utils.flushOut(log)
 
             #Run the job and keep track of the process
-            if sys.platform == "win32":
-                self.childProcess = subprocess.Popen(self.renderCMD,
-                                                    stdout = log,
-                                                    **Utils.buildSubprocessArgs(False))
-            else:
-                self.childProcess = subprocess.Popen(self.renderCMD,
-                                                    stdout = log,
-                                                    **Utils.buildSubprocessArgs(False))
+            self.childProcess = subprocess.Popen(self.renderCMD, stdout = log,
+                                                **Utils.buildSubprocessArgs(False))
 
             logger.info('Started PID {0} to do Task {1}'.format(self.childProcess.pid,
                                                                 render_task.id))
@@ -175,46 +160,41 @@ class RenderTCPServer(TCPServer):
             #Wait until the job is finished or terminated
             self.childProcess.communicate()
 
-            #Gather info on exit and record
-            render_task.exitCode = self.childProcess.returncode
+            #When finished, gather info on exit and record
             if tThread:
                 self.timeoutThread.cancel()
+
+            render_task.exitCode = self.childProcess.returncode
             logString = "\nProcess exited with code {0} at {1} on {2}\n"
-            log.write(logString.format(render_task.exitCode,
-                                        datetime.datetime.now().replace(microsecond=0),
+            nowTime = datetime.datetime.now().replace(microsecond = 0)
+            log.write(logString.format(render_task.exitCode, nowTime,
                                         self.thisNodeName))
 
         except Exception, e:
             traceback.print_exc(e, log)
             raise
 
+        #Finally, update the DB with the information from the task
         finally:
             #Get the latest info about this render node
             with transaction() as t:
                 thisNode = hydra_rendernode.fetch("WHERE host = %s",
-                                                            (self.thisNodeName,),
-                                                            explicitTransaction=t)
+                                                    (self.thisNodeName,),
+                                                    explicitTransaction = t)
 
                 error = False
-                #Check if job was killed, update the job board accordingly
                 if self.childKilled:
-                    #Reset the rendertask
                     render_task.status = self.statusAfterDeath
-                    #render_task.startTime = None
-                    #render_task.host = None
                     self.childKilled = False
-                    #Mark error if task has timedout
                     if self.statusAfterDeath == TIMEOUT:
                         error = True
                 else:
-                    #Report that the job was finished if exit code is 0
                     if render_task.exitCode == 0:
                         render_task.status = FINISHED
                         render_task.endTime = datetime.datetime.now()
-                    #Else, mark error
                     else:
                         error = True
-                #Set as error
+
                 if error:
                     render_task.attempts += 1
                     if render_task.failures == None:
@@ -227,14 +207,9 @@ class RenderTCPServer(TCPServer):
                         render_task.endTime = datetime.datetime.now()
                     else:
                         render_task.status = READY
-                        render_task.host = None
-                        render_task.logFile = None
                         render_task.startTime = None
                         render_task.endTime = None
-                        render_task.exitCode = None
 
-
-                #Return to 'IDLE' IF current status is 'STARTED'
                 if thisNode.status == STARTED:
                     thisNode.status = IDLE
                 elif thisNode.status == PENDING:
@@ -242,67 +217,72 @@ class RenderTCPServer(TCPServer):
                 logger.debug("Node Status: {0}".format(thisNode.status))
                 thisNode.task_id = None
 
-                #Update the records
                 render_task.update(t)
                 thisNode.update(t)
 
             log.close()
-            #Discard info about the previous child process
             self.childProcess = None
-            #Update taskCount
+            self.childKilled = False
+            self.statusAfterDeath = None
             JobUtils.updateJobTaskCount(render_task.job_id)
             JobUtils.manageNodeLimit(render_task.job_id)
             logger.info('Done with render task {0}'.format(render_task.id))
 
     def killCurrentJob(self, statusAfterDeath):
         """Kills the render node's current job if it's running one."""
-        #Get the subprocesses of the main child process
-        try:
-            psutil_proc = psutil.Process(self.childProcess.pid)
-            children_procs = psutil_proc.children(recursive=True)
-        except psutil.NoSuchProcess:
-            children_procs = []
-
-        for proc in children_procs:
-            try:
-                logger.info("Killing subtask with PID of {0}".format(proc.pid))
-                os.kill(proc.pid, signal.SIGTERM)
-            except WindowsError as err:
-                logger.error("Could not kill PID {0} due to a WindowsError")
-                logger.error(str(err))
-
-        try:
-            logger.info("Killing main task with PID of {0}".format(self.childProcess.pid))
-            os.kill(self.childProcess.pid, signal.SIGTERM)
-        except WindowsError as err:
-            logger.error("Could not kill PID {0} due to a WindowsError")
-            logger.error(str(err))
         self.childKilled = True
         self.statusAfterDeath = statusAfterDeath
+        #Gather subprocesses just in case
+        if psutil.pid_exists(self.childProcess.pid):
+            psutil_proc = psutil.Process(self.childProcess.pid)
+            children_procs = psutil_proc.children(recursive=True)
+        else:
+            logger.error("PSUtil was unable to find process with PID {0}".format(self.childProcess.pid))
+            children_procs = []
+
+        #Try to kill the main process
+        try:
+            logger.info("Killing main task with PID {0}".format(self.childProcess.pid))
+            os.kill(self.childProcess.pid, signal.SIGTERM)
+        except WindowsError as err:
+            logger.error("Could not kill PID {0} due to a WindowsError".format(self.childProcess.pid))
+            logger.error(str(err))
+            self.childKilled = False
+
+        #Kill the children if they still exist
+        for proc in children_procs:
+            if psutil.pid_exists(proc.pid):
+                try:
+                    logger.info("Killing subtask with PID of {0}".format(proc.pid))
+                    os.kill(proc.pid, signal.SIGTERM)
+                except WindowsError as err:
+                    #Logging this errors as debug since it probably means they're already dead
+                    logger.debug("Could not kill PID {0} due to a WindowsError".format(proc.pid))
+                    logger.debug(str(err))
+            else:
+                logger.debug("Skipping PID {0} because it is probably already dead.".format(proc.pid))
 
     def timeoutCurrentJob(self):
         logger.warning("Current job is timed out, killing...")
         self.timeoutThread.cancel()
         self.killCurrentJob(TIMEOUT)
-        if self.statusAfterDeath == TIMEOUT:
-            logger.info("Job timed out!")
-        else:
-            logger.info("Job was not timed out!")
+        if self.childKilled:
+            logger.info("Task was timed out")
 
     def startTimeoutThread(self, timeout):
         self.timeoutThread = threading.Timer(timeout, self.timeoutCurrentJob)
         try:
             self.timeoutThread.start()
-            hours = timeout / 60
-            minutes  = timeout % 60
-            infoStr = "Starting Timeout Thread for {0} hours and {1} seconds"
-            logger.info(infoStr.format(hours, minutes))
+            minutes = timeout / 60
+            seconds  = timeout % 60
+            infoStr = "Starting Timeout Thread for {0} minute(s) and {1} second(s)"
+            logger.info(infoStr.format(minutes, seconds))
         except Exception as err:
             self.timeoutThread.cancel()
             logger.error("Could not start timeout thread!")
             logger.error(str(err))
 
-def heartbeat(interval = 5):
+def heartbeat(interval = 60):
     host = Utils.myHostName()
     while True:
         try:
@@ -314,18 +294,18 @@ def heartbeat(interval = 5):
         time.sleep(interval)
 
 def checkRenderNodeInstances():
-    #TODO: Fix This
     if sys.platform == "win32":
         subprocessOutput = subprocess.check_output('tasklist', **Utils.buildSubprocessArgs(False))
         nInstances = len(filter(lambda line: 'RenderNode' in line,
                         subprocessOutput.split('\n')))
     else:
+        subprocessOutput = subprocess.check_output(["ps", "-af"], **Utils.buildSubprocessArgs(False))
         nInstances = len(filter(lambda line: 'RenderNode' in line,
-                        subprocess.check_output(["ps", "-af"], **Utils.buildSubprocessArgs(False)).split('\n')))
+                        subprocessOutput.split('\n')))
     logger.debug("{0} RenderNode instances running.".format(nInstances))
 
-    if nInstances > 2:
-        logger.error("Blocked RenderNodeMain from running because another"
+    if nInstances > 1:
+        logger.critical("Blocked RenderNodeMain from running because another"
                     " instance already exists.")
         return False
